third_party/WebKit/LayoutTests/fast/css-grid-layout/named-grid-line-get-set-expected.txt

Test that setting and getting grid-definition-columns and grid-definition-rows works as expected

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


Test getting grid-definition-columns and grid-definition-rows set through CSS
PASS window.getComputedStyle(gridWithFixedElement, '').getPropertyValue('grid-definition-columns') is "(first) 10px"
PASS window.getComputedStyle(gridWithFixedElement, '').getPropertyValue('grid-definition-rows') is "(first) 15px"
PASS window.getComputedStyle(gridWithPercentElement, '').getPropertyValue('grid-definition-columns') is "424px (last)"
PASS window.getComputedStyle(gridWithPercentElement, '').getPropertyValue('grid-definition-rows') is "162px (last)"
PASS window.getComputedStyle(gridWithPercentWithoutSize, '').getPropertyValue('grid-definition-columns') is "0px (last)"
PASS window.getComputedStyle(gridWithPercentWithoutSize, '').getPropertyValue('grid-definition-rows') is "0px (last)"
PASS window.getComputedStyle(gridWithPercentWithoutSizeWithChildren, '').getPropertyValue('grid-definition-columns') is "77px (last)"
PASS window.getComputedStyle(gridWithPercentWithoutSizeWithChildren, '').getPropertyValue('grid-definition-rows') is "22px (last)"
PASS window.getComputedStyle(gridWithAutoElement, '').getPropertyValue('grid-definition-columns') is "(first) 0px"
PASS window.getComputedStyle(gridWithAutoElement, '').getPropertyValue('grid-definition-rows') is "0px (last)"
PASS window.getComputedStyle(gridWithAutoWithChildrenElement, '').getPropertyValue('grid-definition-columns') is "(first) 77px"
PASS window.getComputedStyle(gridWithAutoWithChildrenElement, '').getPropertyValue('grid-definition-rows') is "22px (last)"
PASS window.getComputedStyle(gridWithMinMax, '').getPropertyValue('grid-definition-columns') is "(first) 80px"
PASS window.getComputedStyle(gridWithMinMax, '').getPropertyValue('grid-definition-rows') is "300px (last)"
PASS window.getComputedStyle(gridWithFixedMultiple, '').getPropertyValue('grid-definition-columns') is "(first nav) 10px (last)"
PASS window.getComputedStyle(gridWithFixedMultiple, '').getPropertyValue('grid-definition-rows') is "(first nav) 15px (last)"
PASS window.getComputedStyle(gridWithPercentageSameStringMultipleTimes, '').getPropertyValue('grid-definition-columns') is "(first nav) 80px (nav) 120px (last)"
PASS window.getComputedStyle(gridWithPercentageSameStringMultipleTimes, '').getPropertyValue('grid-definition-rows') is "(first nav2) 150px (nav2) 450px (last)"
PASS window.getComputedStyle(gridWithRepeatElement, '').getPropertyValue('grid-definition-columns') is "(first) 10px (nav nav2) 400px (nav nav2) 400px"
PASS window.getComputedStyle(gridWithRepeatElement, '').getPropertyValue('grid-definition-rows') is "100px (nav nav2) 150px (nav nav2) 150px (last)"
PASS window.getComputedStyle(gridWithEmptyParentheses, '').getPropertyValue('grid-definition-columns') is "10px"
PASS window.getComputedStyle(gridWithEmptyParentheses, '').getPropertyValue('grid-definition-rows') is "20px 50px"

Test getting and setting grid-definition-columns and grid-definition-rows through JS
PASS getComputedStyle(element, '').getPropertyValue('grid-definition-columns') is "(first) 18px"
PASS element.style.gridDefinitionColumns is "(first) 18px"
PASS getComputedStyle(element, '').getPropertyValue('grid-definition-rows') is "66px (last)"
PASS element.style.gridDefinitionRows is "66px (last)"
PASS getComputedStyle(element, '').getPropertyValue('grid-definition-columns') is "(first) 440px"
PASS element.style.gridDefinitionColumns is "(first) 55%"
PASS getComputedStyle(element, '').getPropertyValue('grid-definition-rows') is "240px (last)"
PASS element.style.gridDefinitionRows is "40% (last)"
PASS getComputedStyle(element, '').getPropertyValue('grid-definition-columns') is "(first) 0px"
PASS element.style.gridDefinitionColumns is "(first) auto"
PASS getComputedStyle(element, '').getPropertyValue('grid-definition-rows') is "0px (last)"
PASS element.style.gridDefinitionRows is "auto (last)"
PASS getComputedStyle(element, '').getPropertyValue('grid-definition-columns') is "(first) 0px"
PASS element.style.gridDefinitionColumns is "(first) min-content"
PASS getComputedStyle(element, '').getPropertyValue('grid-definition-rows') is "0px (last)"
PASS element.style.gridDefinitionRows is "min-content (last)"
PASS getComputedStyle(element, '').getPropertyValue('grid-definition-columns') is "(first) 0px"
PASS element.style.gridDefinitionColumns is "(first) max-content"
PASS getComputedStyle(element, '').getPropertyValue('grid-definition-rows') is "0px (last)"
PASS element.style.gridDefinitionRows is "max-content (last)"
PASS getComputedStyle(element, '').getPropertyValue('grid-definition-columns') is "(first) 440px"
PASS element.style.gridDefinitionColumns is "(first) minmax(55%, 45px)"
PASS getComputedStyle(element, '').getPropertyValue('grid-definition-rows') is "240px (last)"
PASS element.style.gridDefinitionRows is "minmax(30px, 40%) (last)"
PASS getComputedStyle(element, '').getPropertyValue('grid-definition-columns') is "(first) 220px"
PASS element.style.gridDefinitionColumns is "(first) minmax(22em, max-content)"
PASS getComputedStyle(element, '').getPropertyValue('grid-definition-rows') is "50px (last)"
PASS element.style.gridDefinitionRows is "minmax(max-content, 5em) (last)"
PASS getComputedStyle(element, '').getPropertyValue('grid-definition-columns') is "(first) 220px"
PASS element.style.gridDefinitionColumns is "(first) minmax(22em, min-content)"
PASS getComputedStyle(element, '').getPropertyValue('grid-definition-rows') is "50px (last)"
PASS element.style.gridDefinitionRows is "minmax(min-content, 5em) (last)"
PASS getComputedStyle(element, '').getPropertyValue('grid-definition-columns') is "(first) 0px"
PASS element.style.gridDefinitionColumns is "(first) minmax(min-content, max-content)"
PASS getComputedStyle(element, '').getPropertyValue('grid-definition-rows') is "0px (last)"
PASS element.style.gridDefinitionRows is "minmax(max-content, min-content) (last)"
PASS getComputedStyle(element, '').getPropertyValue('grid-definition-columns') is "(first nav) 0px (last)"
PASS element.style.gridDefinitionColumns is "(first nav) minmax(min-content, max-content) (last)"
PASS getComputedStyle(element, '').getPropertyValue('grid-definition-rows') is "(first nav) 0px (last)"
PASS element.style.gridDefinitionRows is "(first nav) minmax(max-content, min-content) (last)"
PASS getComputedStyle(element, '').getPropertyValue('grid-definition-columns') is "(first nav) 0px (nav) 0px (last)"
PASS element.style.gridDefinitionColumns is "(first nav) minmax(min-content, max-content) (nav) auto (last)"
PASS getComputedStyle(element, '').getPropertyValue('grid-definition-rows') is "(first nav2) 0px (nav2) 15px (last)"
PASS element.style.gridDefinitionRows is "(first nav2) minmax(max-content, min-content) (nav2) minmax(10px, 15px) (last)"
PASS getComputedStyle(element, '').getPropertyValue('grid-definition-columns') is "(foo bar) 0px (foo) 0px (bar)"
PASS element.style.gridDefinitionColumns is "(foo bar) auto (foo) auto (bar)"
PASS getComputedStyle(element, '').getPropertyValue('grid-definition-rows') is "(foo bar) 0px (foo) 0px (bar)"
PASS element.style.gridDefinitionRows is "(foo bar) auto (foo) auto (bar)"
PASS getComputedStyle(element, '').getPropertyValue('grid-definition-columns') is "(first) 0px (foo bar) 20px (foo bar) 20px"
PASS element.style.gridDefinitionColumns is "(first) auto (foo bar) 20px (foo bar) 20px"
PASS getComputedStyle(element, '').getPropertyValue('grid-definition-rows') is "220px (foo) 300px (baz)"
PASS element.style.gridDefinitionRows is "220px (foo) 50% (baz)"

Test getting invalid grid-definition-columns and grid-definition-rows set through CSS
PASS window.getComputedStyle(gridWithoutParentheses, '').getPropertyValue('grid-definition-columns') is "none"
PASS window.getComputedStyle(gridWithoutParentheses, '').getPropertyValue('grid-definition-rows') is "none"
PASS window.getComputedStyle(gridWithInvalidNestedParentheses, '').getPropertyValue('grid-definition-columns') is "none"
PASS window.getComputedStyle(gridWithInvalidNestedParentheses, '').getPropertyValue('grid-definition-rows') is "none"
PASS window.getComputedStyle(gridWithUnbalancedParentheses, '').getPropertyValue('grid-definition-columns') is "none"
PASS window.getComputedStyle(gridWithUnbalancedParentheses, '').getPropertyValue('grid-definition-rows') is "none"
PASS window.getComputedStyle(gridWithMisplacedParentheses, '').getPropertyValue('grid-definition-columns') is "none"
PASS window.getComputedStyle(gridWithMisplacedParentheses, '').getPropertyValue('grid-definition-rows') is "none"
PASS window.getComputedStyle(gridWithContiguousParentheses, '').getPropertyValue('grid-definition-columns') is "none"
PASS window.getComputedStyle(gridWithContiguousParentheses, '').getPropertyValue('grid-definition-rows') is "none"

Test getting and setting invalid grid-definition-columns and grid-definition-rows through JS
PASS window.getComputedStyle(element, '').getPropertyValue('grid-definition-columns') is "none"
PASS window.getComputedStyle(element, '').getPropertyValue('grid-definition-rows') is "none"
PASS window.getComputedStyle(element, '').getPropertyValue('grid-definition-columns') is "none"
PASS window.getComputedStyle(element, '').getPropertyValue('grid-definition-rows') is "none"
PASS window.getComputedStyle(element, '').getPropertyValue('grid-definition-columns') is "none"
PASS window.getComputedStyle(element, '').getPropertyValue('grid-definition-rows') is "none"
PASS window.getComputedStyle(element, '').getPropertyValue('grid-definition-columns') is "none"
PASS window.getComputedStyle(element, '').getPropertyValue('grid-definition-rows') is "none"
PASS window.getComputedStyle(element, '').getPropertyValue('grid-definition-columns') is "none"
PASS window.getComputedStyle(element, '').getPropertyValue('grid-definition-rows') is "none"
PASS window.getComputedStyle(element, '').getPropertyValue('grid-definition-columns') is "none"
PASS window.getComputedStyle(element, '').getPropertyValue('grid-definition-rows') is "none"
PASS successfullyParsed is true

TEST COMPLETE
